Put the rest of the linux embed webview details in place
